chore: rename and move jobWizardSlice to add-job
